Moved model classes out of Program.cs Fixed int overflow for SegmentId in Health SDK segment models

+++ src/HealthCloudClient/Models/GolfHoleSegment.cs
@@ -17,7 +17,7 @@ namespace HealthCloudClient
         /// <value>The distance walked by the user for the hole</value>
         public int? DistanceWalked { get; set; }
         /// <value>The unique identifier of the segment</value>
-        public int? SegmentId { get; set; }
+        public long? SegmentId { get; set; }
         /// <value>The start time of the segment</value>
         public DateTime? StartTime { get; set; }
         /// <value>The end time of the segment</value>
